Move extension fitting into separate module.
